requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyJWT==2.8.0
python-multipart==0.0.6
# SECURITY VULNERABILITY (Intentional): Outdated dependency
requests==2.25.1
